requirements.txt
dj-database-url==0.3.0
Django==1.11
django-localflavor==2.1
django-toolbelt==0.0.1
gunicorn==19.3.0
httplib2==0.9.1
psycopg2==2.7
whitenoise==4.0b3
assertpy==0.12
djangorestframework==3.10